backend/requirements.txt
fastapi
uvicorn
python-dotenv
huggingface-hub
PyPDF2
python-multipart
